# Precompile generator templates