--- a/lib/mix/tasks/phx.gen.auth/template_sources/context_functions.ex
+++ b/lib/mix/tasks/phx.gen.auth/template_sources/context_functions.ex
@@ -1,0 +1,7 @@
+defmodule Mix.Tasks.Phx.Gen.Auth.TemplateSources.ContextFunctions do
+  @moduledoc false
+
+  use Mix.Phoenix.TemplateSource, template_patterns: [
+    "priv/templates/phx.gen.auth/context_functions.ex"
+  ]
+end