--- a/lib/mix/tasks/phx.gen.auth/template_sources/test_cases.ex
+++ b/lib/mix/tasks/phx.gen.auth/template_sources/test_cases.ex
@@ -1,0 +1,7 @@
+defmodule Mix.Tasks.Phx.Gen.Auth.TemplateSources.TestCases do
+  @moduledoc false
+
+  use Mix.Phoenix.TemplateSource, template_patterns: [
+    "priv/templates/phx.gen.auth/test_cases.exs"
+  ]
+end