--- a/lib/mix/tasks/phx.gen.json.ex
+++ b/lib/mix/tasks/phx.gen.json.ex
@@ -74,6 +74,7 @@ defmodule Mix.Tasks.Phx.Gen.Json do
   """
 
   use Mix.Task
+  use Mix.Phoenix.TemplateSource, template_patterns: ["priv/templates/phx.gen.json/*.*"]
 
   alias Mix.Phoenix.Context
   alias Mix.Tasks.Phx.Gen
@@ -88,7 +89,7 @@ defmodule Mix.Tasks.Phx.Gen.Json do
     Gen.Context.prompt_for_code_injection(context)
 
     binding = [context: context, schema: schema]
-    paths = Mix.Phoenix.generator_paths()
+    paths = Mix.Phoenix.template_sources()
 
     prompt_for_conflicts(context)