--- a/priv/templates/phx.gen.auth/_menu.html.eex
+++ b/priv/templates/phx.gen.auth/_menu.html.eex
@@ -1,10 +1,10 @@
 <ul>
-<%%= if @current_<%= schema.singular %> do %>
-  <li><%%= @current_<%= schema.singular %>.email %></li>
-  <li><%%= link "Settings", to: Routes.<%= schema.route_helper %>_settings_path(@conn, :edit) %></li>
-  <li><%%= link "Log out", to: Routes.<%= schema.route_helper %>_session_path(@conn, :delete), method: :delete %></li>
+<%%= if @current_<%= @schema.singular %> do %>
+  <li><%%= @current_<%= @schema.singular %>.email %></li>
+  <li><%%= link "Settings", to: Routes.<%= @schema.route_helper %>_settings_path(@conn, :edit) %></li>
+  <li><%%= link "Log out", to: Routes.<%= @schema.route_helper %>_session_path(@conn, :delete), method: :delete %></li>
 <%% else %>
-  <li><%%= link "Register", to: Routes.<%= schema.route_helper %>_registration_path(@conn, :new) %></li>
-  <li><%%= link "Log in", to: Routes.<%= schema.route_helper %>_session_path(@conn, :new) %></li>
+  <li><%%= link "Register", to: Routes.<%= @schema.route_helper %>_registration_path(@conn, :new) %></li>
+  <li><%%= link "Log in", to: Routes.<%= @schema.route_helper %>_session_path(@conn, :new) %></li>
 <%% end %>
 </ul>