--- a/priv/templates/phx.gen.auth/confirmation_new.html.eex
+++ b/priv/templates/phx.gen.auth/confirmation_new.html.eex
@@ -1,6 +1,6 @@
 <h1>Resend confirmation instructions</h1>
 
-<%%= form_for :<%= schema.singular %>, Routes.<%= schema.route_helper %>_confirmation_path(@conn, :create), fn f -> %>
+<%%= form_for :<%= @schema.singular %>, Routes.<%= @schema.route_helper %>_confirmation_path(@conn, :create), fn f -> %>
   <%%= label f, :email %>
   <%%= email_input f, :email, required: true %>
 
@@ -10,6 +10,6 @@
 <%% end %>
 
 <p>
-  <%%= link "Register", to: Routes.<%= schema.route_helper %>_registration_path(@conn, :new) %> |
-  <%%= link "Log in", to: Routes.<%= schema.route_helper %>_session_path(@conn, :new) %>
+  <%%= link "Register", to: Routes.<%= @schema.route_helper %>_registration_path(@conn, :new) %> |
+  <%%= link "Log in", to: Routes.<%= @schema.route_helper %>_session_path(@conn, :new) %>
 </p>