--- a/priv/templates/phx.gen.auth/registration_new.html.eex
+++ b/priv/templates/phx.gen.auth/registration_new.html.eex
@@ -1,6 +1,6 @@
 <h1>Register</h1>
 
-<%%= form_for @changeset, Routes.<%= schema.route_helper %>_registration_path(@conn, :create), fn f -> %>
+<%%= form_for @changeset, Routes.<%= @schema.route_helper %>_registration_path(@conn, :create), fn f -> %>
   <%%= if @changeset.action do %>
     <div class="alert alert-danger">
       <p>Oops, something went wrong! Please check the errors below.</p>
@@ -21,6 +21,6 @@
 <%% end %>
 
 <p>
-  <%%= link "Log in", to: Routes.<%= schema.route_helper %>_session_path(@conn, :new) %> |
-  <%%= link "Forgot your password?", to: Routes.<%= schema.route_helper %>_reset_password_path(@conn, :new) %>
+  <%%= link "Log in", to: Routes.<%= @schema.route_helper %>_session_path(@conn, :new) %> |
+  <%%= link "Forgot your password?", to: Routes.<%= @schema.route_helper %>_reset_password_path(@conn, :new) %>
 </p>