--- a/priv/templates/phx.gen.auth/registration_view.ex
+++ b/priv/templates/phx.gen.auth/registration_view.ex
@@ -1,3 +1,3 @@
-defmodule <%= inspect context.web_module %>.<%= inspect Module.concat(schema.web_namespace, schema.alias) %>RegistrationView do
-  use <%= inspect context.web_module %>, :view
+defmodule <%= inspect @context.web_module %>.<%= inspect Module.concat(@schema.web_namespace, @schema.alias) %>RegistrationView do
+  use <%= inspect @context.web_module %>, :view
 end