… priv/templates/phx.gen.auth/auth_test.exs
… priv/templates/phx.gen.auth/confirmation_controller.ex
… priv/templates/phx.gen.auth/confirmation_controller_test.exs
… priv/templates/phx.gen.auth/conn_case.exs
… priv/templates/phx.gen.auth/context_fixtures_functions.ex
… priv/templates/phx.gen.auth/context_functions.ex
… priv/templates/phx.gen.auth/migration.ex
… priv/templates/phx.gen.auth/notifier.ex
… priv/templates/phx.gen.auth/registration_controller.ex
… priv/templates/phx.gen.auth/registration_controller_test.exs
… priv/templates/phx.gen.auth/reset_password_controller.ex
… priv/templates/phx.gen.auth/reset_password_controller_test.exs
… priv/templates/phx.gen.auth/reset_password_edit.html.eex
… priv/templates/phx.gen.auth/reset_password_new.html.eex
… priv/templates/phx.gen.auth/reset_password_view.ex
… priv/templates/phx.gen.auth/routes.ex
… priv/templates/phx.gen.auth/schema.ex
… priv/templates/phx.gen.auth/schema_token.ex
… priv/templates/phx.gen.auth/session_controller.ex
… priv/templates/phx.gen.auth/session_controller_test.exs
… priv/templates/phx.gen.auth/session_new.html.eex
… priv/templates/phx.gen.auth/session_view.ex
… priv/templates/phx.gen.auth/settings_controller.ex
… priv/templates/phx.gen.auth/settings_controller_test.exs
… priv/templates/phx.gen.auth/settings_edit.html.eex
… priv/templates/phx.gen.auth/settings_view.ex
… priv/templates/phx.gen.auth/test_cases.exs
… priv/templates/phx.gen.channel/channel.ex
… priv/templates/phx.gen.channel/channel_test.exs
… priv/templates/phx.gen.context/access_no_schema.ex