This PR precompiles all of the `phx.gen.*` templates making the generators much faster to run. For example, the `phx.gen.auth` test suite went from 30 seconds to 0.7 seconds. This also adds multiple template sources (`LocalApp`, `PhoenixCompiled` and `Phoenix`) so we can still override templates by placing files in `priv/templates`.

However there are a couple drawbacks with the current implementation
* Compilation time goes from 3 seconds to 13 seconds.
* Any time we change a template, `Mix.Phoenix.TemplateSources.PhoenixCompiled` must recompile all of its templates

I'm trying to figure out if we can improve performance by compiling these templates in parallel. I've compiling the eex with `Task.async_stream/1`, but that doesn't help. I've also tried partitioning the templates into multiple modules (using `:erlang.phash_2` to split them into 4 modules, and then calling `Module.create/3`) in hopes that we can take advantage of Elixir's parallel compiler, but `Module.create/3` seems to create the module sequentially. I'm curious if partitioning would still be good so if one template changed, we would only have to recompile that partition (maybe even so far as 1 module per template).

There's a ton of changes here, so it's broken into 2 commits. The first one switches all of the templates over to use assigns so it's compatible with precompiled templates. The second has the new TemplateSource idea including the precompilation logic.

I'm really curious what approaches I could take to make precompilation more efficient. And if it ends up not making sense, that's cool too. Thanks!