@josevalim 2021-01-27
This is very interesting!

I think part of the reason why compilation is slower is because all templates are defined in a single module. Perhaps it would be best to introduce a macro that embeds chosen templates directly into the task module itself. For example, phx.gen.auth would embed its own templates and so on. Would you be willing to explore these change? In other words, when rendering a precompiled template, you would give the module it is supposed to exist at (which will generally be `__MODULE__`).