@aaronrenner 2021-01-28
That's a good idea! I'll experiment with that approach and report back how it goes.