@aaronrenner 2021-01-29
My latest commit added a macro like you suggested, while still using `Mix.Phoenix.TemplateSources.PhoenixCompiled` to as the common place to dispatch templates to the appropriate modules (could definitely be more efficient and I left some todos).

This helps compilation time a little, but the problem is phx.gen.auth has so many templates including a couple big ones that I still get a warning

```shell-session
$ mix compile
Compiling 70 files (.ex)
Compiling lib/mix/tasks/phx.gen.auth.ex (it's taking more than 10s)
Generated phoenix app
```

It is nice that if I'm working in anything other than phx.gen.auth, I don't get the big compilation hit. Ideally, though, it would only recompile an eex file if it's changed. 

A very hacky way would be to somehow write a file containing a module per eex file to a tmp folder and add that tmp folder to the project's elixirc_paths  (I was reading up on the parallel compiler and it doesn't seem like modules are compiled in parallel, only files). Then we're still left with what would we name these modules and how would we reference them in the code? I was looking at the leex and yecc compiler modules to see if there could be something similar for eex, but still wasn't sure how that would work.

I think I'm really getting off in the weeds here. :sweat_smile: Please let me know if you have any ideas.