@aaronrenner 2021-01-29
On second thought, I could break the slow phx.gen.auth templates into their own modules and have the rest of them in a shared module. I'll try that and report back.