@josevalim 2021-01-29
Yes, that would be my suggestion. Break the phx gen auth templates into two or three modules and that should be good enough. There is likely a good logical way to break them apart too.

but all of this makes me think that Erlang evaluator could be optimized and there are likely low hanging fruits in there.