@aaronrenner 2021-02-02
@josevalim I just pushed a second commit that breaks the templates into separate modules. This brings phoenix compilation time down to 9 seconds. I think there's still some more that can be done to improve compilation speed because touching a template file causes 4 modules to be recompiled, but that could probably be addressed in a future PR.

```shell-session
$ touch priv/templates/phx.gen.auth/registration_controller.ex && mix compile --profile time 
Compiling 4 files (.ex)
[profile] lib/mix/tasks/phx.gen.embedded.ex compiled in 84ms
[profile] lib/mix/tasks/phx.gen.auth/injector.ex compiled in 100ms
[profile] lib/mix/tasks/phx.gen.schema.ex compiled in 159ms
[profile] lib/mix/tasks/phx.gen.auth/template_sources/primary.ex compiled in 2164ms
[profile] Finished compilation cycle of 4 modules in 2164ms
[profile] Finished group pass check of 18 modules in 40ms
```

Let me know what you think!