@josevalim 2021-02-02
Hi @aaronrenner! This looks really good and I think we definitely reached the best we can do here. The compilation times are reasonable to me. However, my last remaining concern is that by having the templates embedded, we are increasing the size of the .beam files (pretty much doubling it) and this will translate to higher memory usage for templates we will never run in production. This is the only aspect I don't know yet how to work around.

Something that I also want to do is to profile the evaluation. Evaluation will always be slower but if it is orders of magnitude slower, it means that are probably some low hanging fruits in there. If I can make the phx.gen.auth tests 3 times faster, then I think it is good enough for us to not merge this.

I will keep you posted.