@josevalim 2021-02-02
Hi @aaronrenner, good news, I could make the tests run ~5x faster with a relatively small change: https://github.com/erlang/otp/pull/3022

If this direction is approved by the Erlang/OTP team, then I propose to not merge this since the next Elixir version with Erlang/OTP 24 should already be considerably faster!