@aaronrenner 2021-02-02
@josevalim Wow, that's a great improvement! Thanks for spending some time and looking into this!

> However, my last remaining concern is that by having the templates embedded, we are increasing the size of the .beam files (pretty much doubling it) and this will translate to higher memory usage for templates we will never run in production. This is the only aspect I don't know yet how to work around.

That's a really good point and I didn't even think of this issue. One way we could solve this is to pull the generators into a separate hex package (like phoenix_view) and only load them during development.

```elixir
  {:phoenix_generators, "~> 1.6", only: :dev}
```

I'd be happy to do this work, but I also understand if it over-complicates things. WDYT?