@josevalim 2021-02-02
@aaronrenner the issue with the generators approach is that often we need to coordinate changes and that will complicate things. So I would hold on for now. My hope is that the eval situation will be addressed, if that doesn't work, then we can discuss how to tackle the last bit. :)