@josevalim 2021-02-03
The OTP PR was accepted, so I think we can close this. I will make sure Elixir v1.12 uses Erlang's optimization. :)